requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.3
PyJWT==2.1.0
gunicorn==20.1.0
django-heroku==0.3.1
drf-spectacular==0.21.0